+++ types.d.ts
@@ -81,7 +81,10 @@ type SettingsType = {
    available_scapers?: Array,
    scrape_interval?: string,
    scrape_delay?: string,
-   version?: string
+   scrape_retry?: boolean,
+   failed_queue?: string[]
+   version?: string,
+   screenshot_key?: string,
 }
 
 type KeywordSCDataChild = {